optimaze search from Usecase
