+++ app/src/main/java/com/edurda77/filmlibrary/data/Movie.kt
@@ -8,6 +8,7 @@ class Movie (
     val movieGanre: String,
     val runtime: Int,
     val popularity: Double,
+
     val release_date: String,
     val budget: Int,
     val revenue: Int,
