+++ app/src/main/java/com/edurda77/filmlibrary/domain/TheMDBRepoUseCac.kt
@@ -2,10 +2,11 @@ package com.edurda77.filmlibrary.domain
 
 import androidx.annotation.WorkerThread
 import com.edurda77.filmlibrary.data.Movie
+import com.edurda77.filmlibrary.data.ResultSearchMovie
 import javax.security.auth.callback.Callback
 
 interface TheMDBRepoUseCace {
     @WorkerThread
-    fun getReposForUserSync (userName: String): List<Movie>
-    fun getReposForUserAsync (userName: String, callback: (List<Movie>)->Unit)
+    fun getReposForUserSync (userName: String): List<ResultSearchMovie>
+    fun getReposForUserAsync (userName: String, callback: (List<ResultSearchMovie>)->Unit)
 }
