+++ app/src/main/java/com/edurda77/filmlibrary/ui/AppSearsh.kt
@@ -2,10 +2,11 @@ package com.edurda77.filmlibrary.ui
 
 import android.app.Application
 import android.content.Context
+import com.edurda77.filmlibrary.data.WebTheMdbRepoUsecaseImpl
 import com.edurda77.filmlibrary.domain.TheMDBRepoUseCace
 
 class AppSearsh : Application() {
-    val theMDBRepoUseCace: TheMDBRepoUseCace by lazy { todo }
+    val theMDBRepoUseCace: TheMDBRepoUseCace by lazy { WebTheMdbRepoUsecaseImpl() }
 }
 
 val Context.app
